Adapt nested locations (#2045)
* add sv_node_utils

* add frame/nested checking for rclick
